+++ test/feature/dummy-feature.js
@@ -6,24 +6,28 @@ const groupsSource = factory.resource('groups.bpmn');
 
 Feature('Dummy', () => {
   Scenario('Group of elements with categories', () => {
+    /** @type {import('bpmn-elements').ContextInstance} */
     let context;
     /** @type {import('bpmn-elements').Definition} */
     let definition;
 
     let ended;
-    When('a source with groups is ran', async () => {
+    Given('a source with groups is ran', async () => {
       context = await testHelpers.context(groupsSource);
       definition = new Definition(context);
+    });
+
+    When('ran', () => {
       ended = definition.waitFor('end');
       definition.run();
     });
 
-    And('category can be retrieved', () => {
-      expect(definition.getElementById('Category_1275ejz')?.placeholder).to.be.true;
-    });
-
     Then('it runs to end', () => {
       return ended;
+    });
+
+    And('category can be retrieved from definition', () => {
+      expect(definition.getElementById('Category_1275ejz').placeholder).to.be.true;
     });
   });
 });